LiveEditBackend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-genai==0.8.0
psycopg2==2.9.11
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
google-ai-generativelanguage==0.4.0
google-api-core==2.29.0
google-auth==2.47.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
proto-plus==1.27.0
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
requests==2.32.5
rsa==4.9.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
paystackapi==2.1.0
celery==5.3.6
redis==5.0.4
moviepy==1.0.3
